# Micro Perf - Store a value for 'trace is enabled'