The 'pipeline progression' function that implements the actual next() call is an extremely hot path, and checks for a Tracer to see if it exists in each call. 

This change is a small optimisation to store a 'TracingEnabled' flag at the start of the request, to speed up that check slightly.

### DeepGraphBenchmark

this PR

|  Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
| Resolve | 17.55 us | 0.221 us | 0.196 us | 4.0894 |     - |     - |  16.71 KB |

v6

|  Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
| Resolve | 17.83 us | 0.265 us | 0.222 us | 4.0894 |     - |     - |  16.71 KB |