@tillig 2020-06-04
>Dango is a Japanese dumpling and sweet made from mochiko, related to mochi. It is often served with green tea. Dango is eaten year-round, but the different varieties are traditionally eaten in given seasons. Three to five dango are often served on a skewer.