@alistairjevans 2020-06-04
I think the unicode consortium team have been given too many bytes to put emojis in...